pyproject.toml
[project]
name = "engligen"
version = "0.1.0"
description = "Gerador de exercícios de Inglês em formato de imagem."
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "pillow",
    # Typer não é mais necessário para o menu, pode ser removido se quiser
    # "typer[all]", 
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
]

# ---- ALTERAÇÃO NECESSÁRIA AQUI ----
[project.scripts]
# Mude de 'engligen.main:app' para 'engligen.main:run'
engligen = "engligen.main:run"